feat(xanim): land upstream dobj/dobj_utils/xanim_calc into POSIX/Switch build
Wire src/xanim/dobj.cpp (1333 lines), src/xanim/dobj_utils.cpp
(583 lines), and src/xanim/xanim_calc.cpp (1897 lines) into both
backends — DObj lifetime/lock/skel queries, hide-part bits, tracelines,
and per-frame XAnim accumulation.

Source fixes for -Werror clean build:
  - Replace 32-bit `static_assert(sizeof(DObj_s) - sizeof(models) == 96)`
    with a guard that only fires on 32-bit hosts; on 64-bit the layout
    grows by 12 bytes (tree/skel.mat/models pointers each gain 4).
  - [[maybe_unused]] on hex-rays scratch locals across the three files
    (DObjTraceline, DObjGetSurfaceData, several XAnim_Calc* helpers,
    DObjCalcAnim, XAnim_CalcPosDeltaEntire).
  - Cast the `Destination/Exchange/Comparand` triple in DObjLock so
    InterlockedCompareExchange resolves with `unsigned int` deductions.

Dup removal:
  - Comment out 44 DObj* / XAnim_CalcDeltaForTime / I_dmaGetDObjSkel
    stubs in posix_backbone_stubs.cpp now provided by real upstream.

Satellite stubs:
  - Add R_GetBaseLodDist / R_GetAdjustedLodDist so dobj_utils links
    without pulling D3D-bound r_state_utils.

Drive-by 32-to-64 fixes (KISAKHACK uintptr_t bridges) in
gfx_d3d/r_dpvs_entity, r_meshdata, rb_spotshadow that surfaced while
chasing dependents.

xanim_readwrite.cpp held back — SP-only (pulls server.h save header).
xanim_load_obj.cpp held back — needs typed struct refactor of the
hex-rays char[1092] scratch with 32-bit pointer offsets.

+++ src/xanim/dobj_utils.cpp
@@ -531,7 +531,7 @@ void __cdecl DObjLock(DObj_s *obj)
     {
         while (*Destination)
             ;
-    } while (InterlockedCompareExchange(Destination, 1, 0));
+    } while (InterlockedCompareExchange(Destination, static_cast<unsigned int>(1), static_cast<unsigned int>(0)));
 }
 
 void __cdecl DObjUnlock(DObj_s *obj)
